Reactivate the original TSC CEGUI theme instead of TaharezLook.
This involved quite a bit of conversion effort, but CEED's
`ceed-migrate' script was a great help. The theme needed some manual
editing afterwards, though, because there was a name clash (the old
theme overwrote CEGUI's own TaharezLook theme) and a new widget was
added in CEGUI 0.8 (Label) that needed to be integrated.

This commit touches a lot of C++ source code files, but the only thing
it does there is replacing the reference to "TaharezLook" with the TSC
theme name ("TSCLook256").

Fixes #646.