Change the way the SSL is loaded.
This commit removes the TSC.require scripting method, because the way
it was implemented was incompatible with mruby 2.0.0. The method
caused a segfault related to trying to load a code string while a code
string was already being loaded (nested loading).

Instead, the entire SSL is now loaded on level startup into memory
right away. That way, there is no need for TSC.require anymore, and it
was an artificial in-between step anyway.

As a consequence, all levels using TSC.require for loading a file from
the SSL need to be changed so that they don't call TSC.require
anymore. As semantically nothing changed, simply removing the
TSC.require statement is enough. This commit does so for all levels
included with TSC itself.

Note: If you already have a copy of this repository, you should run:

  $ git submodule update

Because the last commit changed the mruby submodules.