Add cScene class and MODE_SCENE game mode for visual novel telling.
The cScene class represents a control class for a new game mode,
MODE_SCENE. Entering this mode enters visual novel telling mode, alias
scene mode. In this mode, a single background image is shown at a
time, with text boxes to tell the story (yet to be implemented).

cScene is the superclass for all scenes. This commit adds a single
subclass that's currently still quite empty, cPrologueScene, and which
is intended to describe the game prologue.

Each scene can make use of separate "scene actions", where an action
describes something that is happening in the scene. For instance,
there are/will be actions for changing the background image, showing
text to the user, etc. cScene::Update() executes the list of actions
in a scene one-by-one, stopping on actions that request it, so that
the action gets opportunity to process user input.

This is part of ticket #111.