Use a StaticText instead of MultiLineEditbox for the story box.
Reason: StaticText can parse colours, whilst MultiLineEditbox
can't. Since the user is not supposed to edit the story text while
watching a scene, StaticText is a much better fit.

Finding out about the "HorzFormatting" property wasn't easy,
though. It's not really documented, but it's required for line wrap in
a StaticText.