Improve integration of scenes with other gameplay.
Scenes can now be played both from within a level and from any other
game mode. If a scene is played from within a level, calling
cSceine::Set_Next_Level() is not required; the scene will resume the
pActive_Level just where the level was left off. If on the other hand
a scene is started from outside a level, calling Set_Next_Level() is
obligatory, because otherwise the scene would try to resume
pActive_Level on finish, which would crash. If Set_Next_Level() is
called, the scene hands over to the level specified there
instead. Optionally, a level entry can be specified as well.