Exempt overworld connection lines from MASS_PASSIVE mantissa. Fixes #370.
As soon as the overworld loader came over a line point the Z mantissa
for MASS_PASSIVE (overworld line points the MASS_PASSIVE assigned)
used in cSprite_Manager::Ensure_Different_Z() (for the
cWorld_Sprite_Manager subclass) jumped to 0.87, which is the m_pos_z
for line points. Waypoints on the other hand have a Z position of
0.08 and are also MASS_PASSIVE. When a number of new objects were
added, the Z mantissa for MASS_PASSIVE ultimately increased over 0.08,
causing any new objects to have a higher Z value as the waypoints.

Solution applied by this commit is to exempt the overworld connection
line points from affecting the Z mantissa for MASS_PASSIVE.