Convert Rakefile to Makefile.
All the project's adminsitrative tasks are now to be invoked with make
instead of rake. Make is more universally available, and more people
can deal with writing Makefiles than with writing Rakefiles.

The automatic cross-compilation for the release task has been removed
as cross-compilation support was removed in an earlier commit.