Implement a two-line status bar.
This was required, because a one-line status bar is not large enough
to show all the information that until now has been attached to the
mouse cursor or the HUD, both of which the status bar is meant to
replace.

This commit does not yet remove the mouse cursor and HUD information.