Remove "Editor enabled" message.
With the new status bar, the HUD can stay permanently hidden, making
the change introduced with commit
e93a3ccfa1abe9576d15ae8ec9789a53ba7cd79d permanent. When the HUD
however is hidden, HUD messages do not expire. This caused this
message to be displayed after quitting the editor when the HUD is
shown again. Thus, this commit removes that message.