Implement background preview. Closes #156.
It works by stacking multiple StaticImage CEGUI window's on top of
each other. At the bottom an image using TSCLook256/GenericBrush (a
white pixel) is employed onto which the colour gradient is applied
after the image has been stretched accordingly. Above this comes one
StaticImage for each background to preview.

This result is achived by adding the colour gradient image directly
into the XML, whereas for the background images only an invisible
container window (of the same size as the colour gradient image) has
been added to the XML. The C++ code for the event handlers in the
level settings editor's background tab then adds and deletes the
StaticImage instances to and from the invisible container window
provided by the XML.