Revamp script data save mechanism.
This commit completely restructures the way level scripts can save
data into the savegame file. The old `save' and `load' events are
removed and replaced by a new `save_load' event; this way it is
ensures that there cannot be an unbalanced number of save and load
event handlers.

The limitation of only one save/load handler is lifted. It is now
possible to define an arbitrary number of save/load handlers, as long
as they are defined unconditionally at the top level (otherwise the
load mechanism fails to map the script data to the correct load
handler).

The way how script data is stored in the savegame XML is completely
rewritten. The old mechanism used the mruby-json GEM to recursively
dump a hash into string form as JSON and the JSON was then stored in
the XML. On loading, the JSON was parsed. However, using JSON inside
XML violates good taste. The new mechanism is entirely XML based: all
data is stored as normal XML attributes and nodes. As a side effect,
this allows to drop the dependency on the mruby-json GEM.