Restore Std::Switch's state on savegame load if requested. Closes #579.
This commit adds a new option :save to Std::Switch::new() that, if
set, makes Std::Switch register a handler for a `save_load' event that
saves its state on saving and restores the state on loading a
savegame.

The option also accepts the value :callback, which additionally runs
the switch's callback if the switch is found activated on savegame
loading.