Remove all non-default MGEMs.
None of them were in any active use, and they were cluttering TSC's
repository. The only MGEM that was in use use the JSON one, but as the
savegame mechanism has now been reworked to use XML with a custom
implementation, it is not needed anymore either. Thus, this commit
removes all the Git submodules and the related mruby build
instructions.