Set version to 2.1.1-dev.
The `devel' branch now targets the first post-2.1.0 release.
Development for 2.1.0 occurs in the `release-2.1.0' branch.

The next version after 2.1.0 is probably 2.1.1, so 2.1.1-dev was
chosen as the version. This is not a binding decision and may later be
changed to 2.2.0-dev or whatever best fits the state of the code.